INSTALLDIR, use find, and add bin directory.
dotfiles are installed in $INSTALLDIR, instead of $HOME. The default
value of $INSTALLDIR is $HOME. This makes for easy testing by changing
$INSTALLDIR to a temporary location.

A 'bin' directory is included for installing common scripts in
$INSTALLDIR/bin.

The 'find' command is used to locate all objects for linking in the
'files' and 'bin' directories. This method does not work for the 'dirs'
directory, due to varying directory depths for linking, so linking
directories still have to be manually listed in $ddirs.