Fonts directory, and use find on all directories.
Moved Inconsolata-powerline font from dir/fonts to fonts.

All objects for symbolic links are found with find, so no longer need to
manually specify them.

Description of directory structure:

files:

All files in this directory are symbolically linked to from $INSTALLDIR
with a '.' prefix', with the same directory structure under files
maintained in $INSTALLDIR.

dirs:

All directories listed in the root of this directory are linked to from
$INSTALLDIR with a '.' prefixed to each.

bin:

All files are symbolically linked to from the $INSTALLDIR/bin directory.

fonts:

All directories listed in the root of this directory are linked to from
$INSTALLDIR/.fonts.