Initial commit of MyRequiredProgs
A small utility script I use to check if a program exists. The name of
any missing program(s) will be shown on stdout. As an example, if a
script requires the 'echo' and 'ls' commands ( I wouldn't normally check
for these, this is an EXAMPLE!) I would add the following line at the
beginning of the script:

MyRequiredProgs echo ls || exit 1