Chained keys and open to desktop.
W-Enter: urxvt open on desktop 1
W-t : urxvt open on desktop 1

Chained keys:

W-c b: chromium browser open on desktop 2
W-c c: Open Google calendar in chromium on desktop 3
W-c e: Open Gmail in chromium on desktop 3

For the calendar and gmail apps, xdotool is used to determine if the app
is already running. If the app is already open, the window is activated,
otherwise the app is launched inside chromium.

For the just the chromium browser, a new browser window is opened.