Add menu items and associate images to feh script
A menu item named 'My Apps' will be created, and the
The following menu items will be installed:

  - My Apps
    - VMProgs
      - Excel 2013
      - Harmony Remote
      - Materials Studio 7.0
      - PowerPoint 2013
      - Word 2013
    - GMail
    - Google Calendar
    - Google Drive
    - Pandora Internet Radio
    - urxvt (launch through daemon)

Icons which are freely distributable are included.

Invisible menu items are also installed (userapp-feh_browser_fullscreen,
and userapp-feh_browser_windowed), for which
userapp-feh_browser_windowed is set to be used for viewing image
mimetypes.