MyProfileUpdate / Improve bashrc loading.
The MyProfileUpdate script processes all the files in ~/.bash.d and
creates a single file called 'profile.env.' This new file is sourced
from ~/.bashrc.

Previously, ~/.bashrc processed all the files in ~/.bash.d for each
instance, even though the files rarely change. Taking a page from Gentoo
env-update tool, MyProfileUpdate is run when changes in ~/.bash.d are to
be included. The draw-back to this method is that the user is now
responsible for running MyProfileUpdate, however this should not be
needed very often.