vim: Change where swap files are saved
angular-cli monitors every file in a directory on 'ng serve', which
causes a rebuild every time the vim swap file is modified. This commit
moves all vim swapfiles to the $XDG_CACHE_HOME/vim-swap directory, and
uses a special '//' suffix for vim, which avoids file collisions when
editing more than one file with the same name, but different paths.