Prompt changes (history # and error code)
Do not show history number at the beginning of the prompt since I never
took advantage of this feature, however I did use the color changing of
the history number as an indication of the error code.

The error code is now indicated by a green check mark, or a red 'x' at
the end of the prompt.