Fix prompt for newer git
Somewhere after git 2.16.4, the capitalization of the message received
when running 'git status' in a directory which is not a git repository
changed slightly, from 'fatal: Not a git repository' to 'fatal: not a
git repository'.  This commit fixes the bash prompt command by searching
for '[Nn]ot', so that it will work for both newer and older git
versions.